Adding list of compression commands

+++ src/mydumper_global.h
@@ -127,5 +127,5 @@ extern gboolean it_is_a_consistent_backup;
 extern gchar *partition_regex;
 extern gchar **exec_per_thread_cmd;
 extern const gchar *compress_method;
-extern guint min_rows_per_file;
-extern guint max_rows_per_file;
+extern guint64 min_rows_per_file;
+extern guint64 max_rows_per_file;
